requirements.txt
discord.py
aiosqlite
python-dotenv
datetime